Generalized side_nav to a partial.
The side_nav html that used to be a part of each controller's individual show.html.erb views is now generalized in a partial. It also now has the ability to apply an additional 'selected_a' class to the anchor for the currently selected project (whether in a project or a photo show view). The appropriate functional tests are included.