Corrected an erroneous 'title' validation assumption in the Photo model.
The Photo model has been updated to no longer validate the presence of a title when new photos are added, as there may be cases where a photo might not have a title. The corresponding unit test has been updated to assert that there are no errors given when inserting a photo without a title field.