Fixed a unit test operating under a legacy assumption.
The photo_test.rb unit test tested for the presence of a title in the first Photo entry, a validation that is no longer in place. Fortunately, the test failed when it was run against the current photos fixtures. I also added a test for the presence of an img_url field in the first Photo entry. This new test passes.