Instance method display_photos of Project model unit tested.
This commit contains a unit test for the display_photos method of the Project model and the YAML file it requires (photos.yml). The method returns an array of all of a project's photos in ascending display_order value. The method's unit test asserts that the values of a pre-established array of all of projects(:proj_two) photos (organized by display_order) match exactly to the returned array of the method.