Added a thumb_url column to Photo model and updated Project#show view accordingly.
A thumb_url column was added to the Photo model, along with the appropriate unit tests. The Project#show view was updated to use this field, generating an image tag thumbnail for each photo in the given project. Each thumbnail div is set with the first_thumb_container class at this point. This will be fixed in a future commit.