Fleshed out dynamic Project#show view.
The Project#show view now displays dynamically generated thumbnails that are hyperlinked to their corresponding photo's page. The appropriate 'first_thumb' and 'last thumb' class appendages are activated according to a super-basic algorithm that determines the column of the photo based on the order in which it is processed in the view.